Improve concurrency and safety
Add async/await and improve the overall concurrency and safety of the code. This also enables the use of the new `StrictConcurrency=complete` compiler flag.

This change was made to improve the reliability and performance of the library. It also ensures that the library is compatible with the latest Swift features.