Add Pushed Authorization Request (PAR) support
Adds support for Pushed Authorization Request (PAR) to allow clients to initiate OAuth authorization flows without user interaction. This introduces a new endpoint for clients to push authorization requests and receive a request URI. The authorize endpoint is extended to handle these requests. A new `PARRequestStorage` protocol is introduced to allow for customizable storage of PAR requests. Two implementations, `InMemoryPARStorage` and `RedisPARStorage`, are provided for in-memory and Redis-based storage, respectively. This feature improves the user experience for clients that require automated authorization flows.

Closes #7